Demo/requirements.txt
# Browser Control
playwright==1.48.0

# LLM Integration
google-generativeai>=0.8.3
pillow

# Utilities
python-dotenv==1.0.1

# Browser-Use (alternative approach with better element detection)
browser-use>=0.11.0
langchain-google-genai>=2.0.0

# LangGraph Browser Agent
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-mcp

# FastMCP 3.0 - https://gofastmcp.com
fastmcp>=2.0.0
